upgrade go and fix cicd
add other paths for exclusion rules and make the line length 150 and add the output configuration;
fix and add the lint, ci, build, and test workflows;
remove the sonar token to avoid the usage of the sonar's analyses;
remove the race and shuffle flags from the make test command;
optimize the TestThrottling test;
set a separate statusCh variable to store the peer client and remove the t.Skip() from the TestStatusPubSub test;
optimize the codebase based on the golintci issues;
change DefaultChainID to 8844;
change error string to start with small letter in order to follow the golang conventions;
actions/setup-go must be done after actions/setup-checkout for cache optimizations.
updated the version of the actions.